完善多处 新增悬浮抽屉栏的选择 增强权限 优化首页

+++ resources/views/community-content/comment-data.blade.php
@@ -8,12 +8,16 @@
         <div class="news-content-comment-p">{!! $reply->content !!}</div>
         <div class="action-area">
             <span class="action-need-hover">
-                <a href="{{route('adminCommunityTopicReplyEdit',[$topic->id,$reply->id])}}" target="_blank" class="mdui-btn mdui-btn-icon mdui-ripple mdui-btn-dense mdui-text-color-pink-accent">
-                    <i class="mdui-icon material-icons">edit</i>
-                </a>
-                <button onclick="deleteNewsReply('{{$reply->id}}','{{str_limit(strip_tags($reply->content), $limit = 20, $end = '...')}}')" class="mdui-btn mdui-btn-icon mdui-ripple mdui-btn-dense mdui-text-color-pink-accent">
-                    <i class="mdui-icon material-icons">delete</i>
-                </button>
+                @hasanyrole('Founder|Maintainer')
+                    <a href="{{route('adminCommunityTopicReplyEdit',[$topic->id,$reply->id])}}" target="_blank" class="mdui-btn mdui-btn-icon mdui-ripple mdui-btn-dense mdui-text-color-pink-accent">
+                        <i class="mdui-icon material-icons">edit</i>
+                    </a>
+                @endhasanyrole
+                @can('update',$reply)
+                    <button onclick="deleteNewsReply('{{$reply->id}}','{{str_limit(strip_tags($reply->content), $limit = 20, $end = '...')}}')" class="mdui-btn mdui-btn-icon mdui-ripple mdui-btn-dense mdui-text-color-pink-accent">
+                        <i class="mdui-icon material-icons">delete</i>
+                    </button>
+                @endcan
                 <a onclick="replyToReply('{{$reply->user->name}}','{{$reply->user->id}}')" class="mdui-btn mdui-btn-dense news-content-comment-reply-btn" ><i class="mdui-icon material-icons mdui-icon-left ">comment</i>回复</a>
             </span>
             <a onclick="ajaxHandleReplyVote('{{route('communityTopicReplyVote')}}','{{route('communityTopicReplyCancelVote')}}','{{$reply->id}}',this)"
